requirements.txt
torch==1.13.0
torchvision==0.14.0
transformers==4.21.1
matplotlib==3.6.0
ipywidgets==7.6.5
ipympl==0.9.2
Pillow
numpy